+++ include/backend/webpage_parsing.h
@@ -16,7 +16,7 @@ typedef struct ContentItem {
 
 typedef struct ContentList {
     ContentItem items[MAX_CONTENT_ITEMS];
-    int num_items;
+    size_t num_items;
 } ContentList;
 
 ParseWebpageContentError parse_stackoverflow_content(char* content, ContentList* content_list);
